chore: move existing tests
