+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "start": "node -r dotenv/config ./prod/index.js",
     "lint": "eslint src --max-warnings 0",
-    "test": "ts-mocha -u tdd test/**/*.spec.ts --recursive --exit"
+    "test": "ts-mocha -u tdd 'test/**/*.spec.ts' --recursive --exit --timeout 5000"
   },
   "repository": {
     "type": "git",
